Move more tests to dedicated files
