+++ tests/Fixtures/Autoloaded/PublicVisibility/pv-non-same-method-parent-m0.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace NonSameAbstract;
+
+abstract class NonSameAbstract
+{
+    public abstract function foo();
+}
+class Child extends NonSameAbstract
+{
+    public function foo()
+    {
+    }
+    protected function bar()
+    {
+    }
+}
