+++ tests/Fixtures/Autoloaded/PublicVisibility/pv-same-method-grandparent-m0.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace SameGrandParent;
+
+class GrandParent
+{
+    protected function foo()
+    {
+    }
+}
+class SameParent extends GrandParent
+{
+}
+class Child extends SameParent
+{
+    public function foo()
+    {
+    }
+}
